backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-socketio>=5.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
motor>=3.3.0
pymongo[srv]>=4.6.0
dnspython>=2.4.0
python-dotenv>=1.0.0
joblib>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.0
certifi>=2024.0.0
gunicorn>=21.2.0
